+++ src/Firebase/Auth/Providers/AuthProvider.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Kreait\Firebase\Auth\Providers;
+
+interface AuthProvider
+{
+    public function getProviderId();
+    public function setProviderId(string $providerId);
+}
