+++ src/Firebase/Auth/Providers/EmailAuthProvider.php
