+++ src/Firebase/Auth/Providers/FacebookAuthProvider.php
