+++ src/Firebase/Auth/Providers/GithubAuthProvider.php
