+++ src/Firebase/Auth/Providers/GoogleAuthProvider.php
