+++ src/Firebase/Auth/Providers/TwitterAuthProvider.php
